Restructure + Rename of some Components,
Removed unnessacary components HighsSolver, GlpkSolver
