--- src/components/HighsSolver.js
@@ -1,47 +1,0 @@
-import React, { useState } from "react";
-import View from "./OutputUi";
-import { solve, SolverOptions } from "../services/SolverInterface";
-
-export const solveHighsProblem = async (inputData, inputFormat, solver) => {
-  try {
-    console.log("SOLVER: " + solver);
-    const startTime = performance.now();
-    const result = await solve(inputData, inputFormat, SolverOptions[solver]);
-    const endTime = performance.now();
-    result["Walltime"] = (endTime - startTime) / 1000;
-
-    return result; // Rückgabe des Ergebnisses
-  } catch (error) {
-    return { error: error.message };
-  }
-};
-
-const HighsSolver = () => {
-  // const [inputData, setInputData] = useState("");
-  // const [outputData, setOutputData] = useState(null);
-  // const [inputFormat, setInputFormat] = useState("GMPL");
-  //
-  // const solveProblem = async (inputData, inputFormat) => {
-  //     const startTime = performance.now();
-  //     const result = await solve(inputData, inputFormat, SolverOptions.HIGHS);
-  //     const endTime = performance.now();
-  //     result["Walltime"] = (endTime - startTime) / 1000;
-  //
-  //   setOutputData(result); // Store the result as an object
-
-  return (
-    <React.Fragment>
-      <div>ferf</div>
-    </React.Fragment>
-    // <View
-    //   inputFormat={inputFormat}
-    //   setInputFormat={setInputFormat}
-    //   inputData={inputData}
-    //   setInputData={setInputData}
-    //   solveProblem={solveProblem}
-    //   outputData={outputData} // Pass the parsed result object to View
-    // />
-  );
-};
-
-export default HighsSolver;
