+++ src/components/Tabs/styles/navbar.css
